requirements.txt
flask
requests
python-dotenv
langdetect
